Create UNIX_date_command
Using the date command , write a complete bash script that parses the current hour of the day and outputs either "YES" (current hour is 8:00 pm or greater) or "NO"(current hour is prior to 8:00pm). Assume the UNIX date command outputs the hour in military format(i.e. 0-23) and displays in the UTC time zone ( which is what you want).